+++ data/examples/truss_zeroforce.obj
