--- data/paper/exA_tree.obj
@@ -1,115 +1,0 @@
-# Rhino
-
-mtllib exA_tree.mtl
-usemtl diffuse_0_0_255_255
-v 2.6619895384620569 1 0
-v 1 -1 0
-cstype bspline
-deg 1
-curv 0 2.8284271247461903 1 2
-parm u 0 0 2.8284271247461903 2.8284271247461903
-end
-usemtl diffuse_0_0_255_255
-v 2.6619895384620569 1 0
-v 4 2 0
-cstype bspline
-deg 1
-curv 0 1.4142135623730951 3 4
-parm u 0 0 1.4142135623730951 1.4142135623730951
-end
-usemtl diffuse_0_0_255_255
-v 2.6619895384620569 1 0
-v 2 2 0
-cstype bspline
-deg 1
-curv 0 1.4142135623730951 5 6
-parm u 0 0 1.4142135623730951 1.4142135623730951
-end
-usemtl diffuse_0_0_255_255
-v 2 2 0
-v 2 3 0
-cstype bspline
-deg 1
-curv 0 1 7 8
-parm u 0 0 1 1
-end
-usemtl diffuse_0_0_255_255
-v 4 2 0
-v 4 3 0
-cstype bspline
-deg 1
-curv 0 1 9 10
-parm u 0 0 1 1
-end
-usemtl diffuse_0_0_255_255
-v 4 2 0
-v 2 2 0
-cstype bspline
-deg 1
-curv 0 2 11 12
-parm u 0 0 2 2
-end
-usemtl diffuse_0_0_255_255
-v -0.66198953846205688 1 0
-v 1 -1 0
-cstype bspline
-deg 1
-curv 0 2.8284271247461903 13 14
-parm u 0 0 2.8284271247461903 2.8284271247461903
-end
-usemtl diffuse_0_0_255_255
-v -0.66198953846205688 1 0
-v -2 2 0
-cstype bspline
-deg 1
-curv 0 1.4142135623730951 15 16
-parm u 0 0 1.4142135623730951 1.4142135623730951
-end
-usemtl diffuse_0_0_255_255
-v -0.66198953846205688 1 0
-v 0 2 0
-cstype bspline
-deg 1
-curv 0 1.4142135623730951 17 18
-parm u 0 0 1.4142135623730951 1.4142135623730951
-end
-usemtl diffuse_0_0_255_255
-v 1 -1 0
-v 1 -2 0
-cstype bspline
-deg 1
-curv 0 1 19 20
-parm u 0 0 1 1
-end
-usemtl diffuse_0_0_255_255
-v 0 2 0
-v 0 3 0
-cstype bspline
-deg 1
-curv 0 1 21 22
-parm u 0 0 1 1
-end
-usemtl diffuse_0_0_255_255
-v -2 2 0
-v -2 3 0
-cstype bspline
-deg 1
-curv 0 1 23 24
-parm u 0 0 1 1
-end
-usemtl diffuse_0_0_255_255
-v 0 2 0
-v 2 2 0
-cstype bspline
-deg 1
-curv 0 2 25 26
-parm u 0 0 2 2
-end
-usemtl diffuse_0_0_255_255
-v -2 2 0
-v 0 2 0
-cstype bspline
-deg 1
-curv 0 2 27 28
-parm u 0 0 2 2
-end
