+++ data/paper/exD_truss.obj
