+++ scripts/example_truss_zeroforce_constraints.py
